+++ presentation/src/main/java/com/kk/presentation/host/progressgame/PreStartAdminView.kt
@@ -1,0 +1,31 @@
+package com.kk.presentation.host.progressgame
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.kk.designsystem.components.*
+import com.kk.presentation.R
+
+@Composable
+fun PreStartAdminView(
+    navigateToWaitingView : () -> Unit
+){
+    //viewmodel round: String
+    val round = "1"
+    Column(modifier = Modifier.fillMaxSize()) {
+        Column(modifier = Modifier
+            .fillMaxWidth()
+            .padding(30.dp), horizontalAlignment = Alignment.CenterHorizontally) {
+            KkTitle(label = round)
+        }
+        Box(modifier = Modifier.padding(25.dp)) {
+            KkBody(label = stringResource(id = R.string.question_to_ask))
+        }
+        Column(modifier = Modifier.fillMaxSize(), horizontalAlignment = Alignment.CenterHorizontally, verticalArrangement = Arrangement.Center) {
+            KkOrangeTitle(label = stringResource(id = R.string.start_round))
+        }
+    }
+}
